+++ NEWS.md
@@ -1,5 +1,7 @@
 # automerge (development version)
 
+* Functions accepting a `heads` parameter now support multiple heads.
+
 * Fixes additional Valgrind false positive in vendored automerge library.
 
 # automerge 0.3.1